python/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
auto-py-to-exe==2.46.0
babel==2.17.0
blinker==1.9.0
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==45.0.1
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.1
Eel==0.18.1
Flask==3.1.1
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
pyperclip==1.9.0
python-dotenv==1.1.0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==80.4.0
sqlparse==0.5.3
tkcalendar==1.6.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
